pyproject.toml
[project]
name = "wings"
version = "0.1.4"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=48.0.0",
    "django>=6.0.5",
    "django-allauth>=65.16.1",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]

[dependency-groups]
dev = [
    "faker>=40.15.0",
    "model-bakery>=1.23.4",
    "pytest-django>=4.12.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "wings.settings"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = ["tests"]
